% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pingRMSCGLR.R
\name{ping.RMSCGLR}
\alias{ping.RMSCGLR}
\title{Projected Iterated Normed Gradient}
\usage{
ping.RMSCGLR(Z, X, AX, W, F, u, method, poids, E)
}
\arguments{
\item{Z}{matrix of the working variables}

\item{X}{matrix of the normalized covariates}

\item{AX}{matrix of the suplementary covariates}

\item{W}{matrix of weights}

\item{F}{matrix of components}

\item{u}{loading vector}

\item{method}{structural relevance criterion}

\item{poids}{weights on the working variables}

\item{E}{set of components}
}
\value{
the updated loading vectors
}
\description{
Projected Iterated Normed Gradient
}
