% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_RMSCGLR}
\alias{plot_RMSCGLR}
\title{rmSCGLR generic plot}
\usage{
plot_RMSCGLR(x, thresold = 0, group = 1, plan = c(1, 2), lin.pred = FALSE)
}
\arguments{
\item{x}{an object from rmSCGLR}

\item{thresold}{correlations with the two components of the plane lower than this threshold will be ignored}

\item{group}{a integer indicating the group}

\item{plan}{a size-2 vector indicating which components are plotted}

\item{lin.pred}{a logical. Should we draw linear predictor}
}
\value{
an object of class \code{\link{ggplot2}}
}
\description{
rmSCGLR generic plot
}
\examples{
\dontrun{
library(rmSCGLR)

# load sample data
data <- genus
data <- as.data.frame(apply(data, 2, as.numeric ))

# get variable names from dataset
n <- names(data)
ny <- n[grep("^gen",n)]
nx <- n[-grep("^gen",n)]
na <- c("geology")
nx <- nx[!nx\%in\%c("geology", "surface")]

# build multivariate formula
form <- multivariateFormula(Y = ny, X = nx, A = na)

# define family
fam <- rep("poisson", length(ny))

# run function
H <- c(2,2)
met <- methodSR.RMSCGLR(l=4, s=0.1, t=0.5)
res <- ResponseMixtureSCGLR(formula=form, data=data, H=H,
                           family=fam, method=met, offset = data$surface)

# plot the results
plot_RMSCGLR(x=res, thresold=0.5, group=1, plan=c(1,2))
}

}
