% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSCGLR.R
\name{ResponseMixtureSCGLR}
\alias{ResponseMixtureSCGLR}
\title{Function that fits the response mixture scglr model}
\usage{
ResponseMixtureSCGLR(
  formula,
  data,
  H = c(2, 2),
  family,
  size = NULL,
  offset = NULL,
  subset = NULL,
  crit = list(),
  method = methodSR.RMSCGLR()
)
}
\arguments{
\item{formula}{an object of class \code{MultivariateFormula} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame to be modeled.}

\item{H}{a vector of interger representing the number of components per group}

\item{family}{a vector of character of the same length as the number of dependent variables:
"bernoulli", "binomial", "poisson" or "gaussian" is allowed.}

\item{size}{describes the number of trials for the binomial dependent variables.}

\item{offset}{used for the poisson dependent variables.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{crit}{a list of two elements : maxit and tol, describing respectively the maximum number of iterations and
the tolerance convergence criterion for the Expectation Maximization algorithm. Default is set to 50 and 10e-6 respectively.}

\item{method}{structural relevance criterion. Object of class "method.RMSCGLR"}
}
\value{
\item{U}{the set of loading vectors}

\item{comp}{the set of components}

\item{eta}{the fitted linear predictor for each group}

\item{coef}{contains regression parameters : intercept, gamma and delta}

\item{pg}{the prior probability of belonging to the group}

\item{cluster}{a vector of integers indicating the cluster to which each response is allocated}

\item{postProbs}{the posterior group membership probabilities}

\item{Y}{the response matrix}

\item{X}{the standardized explanatory matrix}

\item{A}{the supplementary explanatory variables}
}
\description{
Function that fits the response mixture scglr model
}
\examples{
\dontrun{

library(rmSCGLR)

# load sample data
data <- genus
data <- as.data.frame(apply(data, 2, as.numeric ))

# get variable names from dataset
n <- names(data)
ny <- n[grep("^gen",n)]
nx <- n[-grep("^gen",n)]
na <- c("geology")
nx <- nx[!nx\%in\%c("geology", "surface", "forest")]

# build multivariate formula
form <- multivariateFormula(Y = ny, X = nx, A = na)

# define family
fam <- rep("poisson", length(ny))

# run function
H <- c(2,2)
met <- methodSR.RMSCGLR(l=4, s=0.1, t=0.4)
res <- ResponseMixtureSCGLR(formula=form, data=data, H=H,
                           family=fam, method=met, offset = data$surface)
}

}
