% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumlogs.R
\name{sumlogs}
\alias{sumlogs}
\title{Function that computes sumlogs}
\usage{
sumlogs(m)
}
\arguments{
\item{m}{a vector}
}
\value{
the result of sumlogs
}
\description{
Function that computes sumlogs
}
\examples{
\dontrun{
library(rmSCGLR)

sumlogs(c(1,2,3))

}
}
